move pending transactions fetching to `walletStore`
